% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordr.R
\name{addPlots2}
\alias{addPlots2}
\title{Overwrite of addPlots in WordR package, to not crash if a plot isn't in the docx. see WordR::addPlots for details}
\usage{
addPlots2(docxIn, docxOut, Plots = list(), width = 6, height = 6,
  res = 300, style = NULL, debug = F, ...)
}
\description{
Overwrite of addPlots in WordR package, to not crash if a plot isn't in the docx. see WordR::addPlots for details
}
