% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear-all-df-labels.R
\name{clear_all_df_labels}
\alias{clear_all_df_labels}
\title{Remove all labels from a df}
\usage{
clear_all_df_labels(df)
}
\arguments{
\item{df}{df to remove labels from}
}
\value{
df with labels removed
}
\description{
e.g. after using a function in Hmisc that tacks on labels, or
haven::read_dta()
}

