% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{confirm_ids}
\alias{confirm_ids}
\title{Confirm if the rows of data are uniquely identified by cols}
\usage{
confirm_ids(data, cols)
}
\arguments{
\item{data}{data.frame to check}

\item{cols}{char vect of identifier cols}
}
\description{
Confirm if the rows of data are uniquely identified by cols
}
