% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_midpoints.R
\name{cut_midpoints}
\alias{cut_midpoints}
\title{Cut number, return midpoints}
\usage{
cut_midpoints(x, n, ...)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of intervals to create}

\item{...}{Arguments passed on to base::cut.default}
}
\description{
Like ggplot2::cut_number, except returns the midpoints of the intervals rather than intervals (factors).
}
