% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-selected-source.R
\name{dbExecute_selected_sqlite}
\alias{dbExecute_selected_sqlite}
\title{Execute SQLite}
\usage{
dbExecute_selected_sqlite()
}
\description{
Call this function as an rstudio addin to push to a con with dbExecute
}
\details{
note that you cannot use double quotes:
http://stackoverflow.com/questions/1992314/what-is-the-difference-between-single-and-double-quotes-in-sql
}
