% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismisc-extrapolate.R
\name{dismisc_extrapolate}
\alias{dismisc_extrapolate}
\title{Extrapolate (and interpolate) values}
\usage{
dismisc_extrapolate(x, y, xout = x, method = "natural")
}
\arguments{
\item{x}{the independent variable; something like a time value}

\item{y}{the dependent variable; something like a population value}

\item{xout}{values to interpolate/extrapolate on, defaults to x}

\item{method}{can be "lm" or any method in stats::spline()}
}
\value{
the values of the extrapolated function y(x) on domain xout
}
\description{
Uses a spline or lm to extrapolate (and interpolate) values.
}
\details{
Use in a dplyr chain something like this to extrapolate (and interpolate) within city (for example):

\code{

df \%<>\% group_by(city) \%>\% mutate_at(total_population,

funs(population_interpolated = dismisc_extrapolate(x=year,y=.)))
}

If there are no non-missing y values, a vector of NAs is returned.
}

