% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-them-all.R
\name{download_them_all}
\alias{download_them_all}
\title{Download all files in a list of URLs into the working directory}
\usage{
download_them_all(url.vect, skip.already.downloaded = TRUE,
  stub.vect = rep("", length(linkvect)))
}
\arguments{
\item{url.vect}{vector of URLs. Files downloaded will be named according to part of the URL after the last forward slash '/'.}

\item{skip.already.downloaded}{whether to skip files that already exist. If FALSE, they are overwritten.}

\item{stub.vect}{vector of strings to paste0 to the front of destfile names, component-by-component. Use if the part after the last forward slash in the URLs is not unique.}
}
\value{
list of successfully downloaded destfile names
}
\description{
Download all files in a list of URLs into the working directory
}
