% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordr.R
\name{flextable_convert_standard}
\alias{flextable_convert_standard}
\title{Standard conversion to flextables}
\usage{
flextable_convert_standard(out, nblankheaders = 1)
}
\arguments{
\item{out}{A dataframe, columns need to be C1-CN, actual headers in first row.}

\item{nblankheaders}{The number of columns without parentheses numbers over them; default 1.}
}
\description{
Standard conversion to flextables
}
