% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb-get.R
\name{mdb_get}
\alias{mdb_get}
\title{Read in an mdb file.}
\usage{
mdb_get(file, tables = NULL, lowernames = FALSE, mdbexportArgs = "", ...)
}
\arguments{
\item{file}{mdb file to read. You may need to surround the name in quotes, like '"file name.mdb"'.}

\item{tables}{either a list or vector of table names; tables = TRUE to import a list of tables; or the default which means import all tables}

\item{mdbexportArgs}{arguments to pass to mdb-export. See man mdb-export.}

\item{...}{arguments passed to readr::read_csv.}
}
\value{
if only 1 table is read, a tibble. Otherwise, a list of tibbles containing all the tables read in.
}
\description{
A shortened version of Hmisc::mdb.get
}
\details{
Still uses mdbtools internally (so that needs to be installed). On Ubuntu, use sudo apt get install mdbtools.

The process is to use mdbtools to convert the mdb table into a csv file, then use readr::read_csv to read the csv file in quickly.
}

