% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-them-all.R
\name{read_stack}
\alias{read_stack}
\title{Read in and stack a vector of data frames}
\usage{
read_stack(flist, readfnc = readRDS, verbose = FALSE, ...)
}
\arguments{
\item{flist}{Character vector of files (in current working directory)}

\item{readfnc}{Function to use to read in the files in the list (default is readRDS)}

\item{verbose}{Print a message on each iteration}

\item{...}{additional parameters passed to readfnc}
}
\value{
The bind_row()ed dataset
}
\description{
Read in a vector of files, that become data frames, stacking them together with bind_rows() the most memory-efficient way possible.
}
