% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd2rt.R
\name{setwd2rt}
\alias{setwd2rt}
\title{Go to an saved root directory.}
\usage{
setwd2rt(name = "default", add = NULL, list = FALSE)
}
\arguments{
\item{name}{Name of the root to go to, or to add. defaults to 'default'.}

\item{add}{directory to add}

\item{list}{if TRUE, directory is unchanged; a listing of all added roots is printed.}
}
\description{
Basic usage: use setwd2rt(add='rootdir') to set the custom root; use setwd2rt() to go to the custom root; the root is saved as an option (see below), so will be preserved after an e.g. remove(list=ls()). Advanced usage: use setwd2rt(name='rootname',add='rootdir') to add the key-value pair ('rootname','rootdir') to the list of all rootdirs stored; use setwd2rt(name='rootname') to go to 'rootdir'; use setwd2rt(list=TRUE) to list all such key-value pairs.
}
\details{
Best way to use this function

\itemize{
  \item{in your init.R file that you run when you start working on a project}
    \itemize{
      \item{Use setwd2rt(add='rootdir') to specify the root folder of your data project as 'rootdir'}
      \item{Use setwd2rt(add='dependencyDir',name ="someDependency") to specify the root folder of some dependency (e.g. the data from another project that you are using in this project)}
    }
  \item{in your scripts, when you need to get to one of those roots}
   \itemize{
     \item{Use setwd2rt() to go to the 'default' root directory, defined earlier},
     \item{Use setwd2rt(name = "someDependency") to go to the subdir of that dependency defined earlier}
   }
   \item{every collaborator working on your project (e.g. synced via dropbox) should have his/her own init-collaborator-name.R file. This allows different collaborators to have different directory structures.}
}
}
