% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{suup}
\alias{suup}
\title{like -su- in stata, but ONLY for vectors}
\usage{
suup(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
nothing (prints summary output)
}
\description{
like -su- in stata, but ONLY for vectors
}
