% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texhelp.R
\name{t_pstar}
\alias{t_pstar}
\alias{t_parw}
\title{TexHelp functions to clean up regression tables}
\usage{
t_pstar(x)

t_parw(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a "cleaned up" character version of x
}
\description{
t_pstar returns latex encoded starrage, given p-value; t_parw wraps the input with parentheses ().
}
\details{
t_pstar uses "economists" p-values: 0.10 gives one star, 0.05 gives 2 (and the right to publish), 0.01 gives 3.
}
\examples{
t_pstar(c(0.03,0.34,0.06))
# combine with t_td for special effects
t_parw(t_td(c(1.21,3.23,3.2232),d=2))
}
