% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{wrapper for statar::tab, to handle vector input}
\usage{
tab(x, ...)
}
\arguments{
\item{x}{Either a vector or a data.frame}

\item{...}{if x is a data.frame, then a list of variables to tab}
}
\value{
See ?statar::tab
}
\description{
wrapper for statar::tab, to handle vector input
}
