% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_case_when_block.R
\name{to_case_when_block}
\alias{to_case_when_block}
\title{Build case_when block for a mutate()}
\usage{
to_case_when_block(uniqvals, varname)
}
\arguments{
\item{uniqvals}{unique values of varname}

\item{varname}{variable name for the LHS of the case_when block}
}
\value{
prints R code to the console that can be copy and pasted
}
\description{
Build case_when block for a mutate()
}
