% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{like tab from statar, but ONLY works for vectors}
\usage{
tab(x, sf = TRUE, sv = FALSE, ss = 10, sort_freq = sf,
  sort_values = sv, suppress_small = ss)
}
\arguments{
\item{x}{A vector}

\item{sf}{shorthand for sort_freq}

\item{sv}{shorthand for sort_values}

\item{ss}{shorthand for suppress_small}

\item{sort_freq}{sort by frequency large to small.}

\item{sort_values}{sort by values large to small. overrides sort_freq.}

\item{suppress_small}{suppresses beyond this many levels for printing. only applies if sort_freq=TRUE. default 10. set to FALSE or Inf to not suppress.}
}
\value{
a data.frame with the tabulation
}
\description{
From https://github.com/matthieugomez/statar/ with significant modifications
}
